books/dvipdfm.def added to fix missing file in latex builds
texlive has changed the dvipdfm in an incompatible way to handle
Unicode in pdf files. The texlive distribution omits dvipdfm.def
and we need this file. It was added to the books directory and is
copied to the ${MNT}/${SYS}/doc directory during builds. In the
long term we need to change to use dvipdfmx but this cannot be
done now and still maintain backward compatibility.