books/bookvolbib add Special Topics / Polynomial GCD
This introduces a new section of the bibliography on special topics
where there are references with summary paragraphs.

This introduces a section on the special topic of polynomial GCDs and
related papers.

As a side-effect, changepage.sty was introduced into the bookheader.tex
file to make the adjustwidth environment available to the bibliography.

The bookheader.tex file is included in all of the books to make sure
they present a uniform style and function.