books/ProvingAxiomCorrect.pamphlet
I've heard Simon Peyton-Jones (Haskell) give this talk
"How to Write a Great Research Paper"
http://www.youtube.com/watch?v=g3dkRsTqdDA

So, rather stupidly I suppose, I have created this new pamphlet
(books/ProvingAxiomCorrect.pamphlet) which will be an ongoing
public struggle with the question of proving Axiom correct.

At no point in the reasonable future should this be considered
as anything but an attempt to "swim the rope across the river in
order to allow others to build the bridge". In other words, I
don't know how to do this but it needs to be done so...