books/bookvol4, bookvol10.3 What is the purpose of the HACKPI domain?
What is the purpose of the HACKPI domain?

HACKPI is a hack provided for the benefit of the axiom interpreter.
As a mathematical type, it is the simple transcendental extension
Q(\pi) of the rational numbers. This type allows interactive users
to use the name '%pi' without a type both where a numerical value is expected
[ as in draw(sin x,x=-%pi..%pi) ] or when the exact symbolic value is meant.
The interpreter defaults a typeless %pi to HACKPI and then uses the various
conversions to cast it further as required by the context.

One could argue that it is unfair to single %pi out from other constants,
but it occurs frequently enough in school examples (specially for graphs)
so it was worth a special hack. In a non-interactive environment (library),
HACKPI would not exist.

(Manuel Bronstein)