books/bookvol10.3 add information to SingleInteger
This type represents finite (machine) precision integers which wrap
around at a machine dependent value. However, {\bf SingleInteger} has
type {\bf EntireRing} which should have no non-zero divisors. But
\[ (2**16:SINT)**2 = 0\]
on a 32 bit machine. Using {\bf SingleInteger} is a pragmatic
efficiency choice. In this condition, Axiom will complain that the
result is not of type FIXNUM, which is the Common Lisp type for
machine integers.