src/axiom-website/documentation.html add Socrates quote
Then anyone who leaves behind him a written manual, and likewise
anyone who takes it over from him, on the supposition that such
writing will provide something reliable and permanent, must be
exceedingly simple-minded; he must really be ignorant of Ammon's
utterance, if he imagines that written words can do anything more than
remind one who knows that which the writing is concerned with.
  -- Socrates

All I need in the documenation is to be reminded of that which
the writing is concerned with.
 -- Tim Daly