buglist: bug 7268: Typechecker gets confused by flow control
Spad typechecker is supposed to track type tests, and
allow operation which are safe due to dynamic tests.
However, tis works for 'if' but fails for loops, as in:

a:R
b:R
c:=a exquo b
while c case R repeat
  a:=c
  c:=a exquo b

Spad compiler report type error on assignment to a.

But both of these work:

a:R
b:R
c:=a exquo b
if c case R then
  while c case R repeat
    a:=c
    c:=a exquo b

and

a:R
b:R
c:=a exquo b
while c case R repeat
  if c case R then
    a:=c
    c:=a exquo b