projects: progress report for week ending 12/13/14
It has been a busy week. Progress was made on a lot of the projects
listed in the projects file.

-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Axiom continuous integration
-=-=-=-=-=-=-=-=-=-=-=-=-=-=

Most of the work has concentrated on the Axiom continuous integration.
In particular, a effort is being made to clean up all failing
regression test files so that a build is always clean. This is
important to continuous integration since any failure should stop
a deployment.

-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Axiom help files
-=-=-=-=-=-=-=-=-=-=-=-=-=-=

The help documentation for system commands was updated, both in the
Jenks book and Volume 5, the Interpreter. Some existing but undocumented
system commands were documented.

-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Axiom src/input dispersal
-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Along with this effort, existing files in src/input are being reworked.

Some of the files are being folded into the algebra books either as
documentation or unit test cases. The graphics files have been
collected and are being moved to Volume 8.1, the Axiom gallery.

Integration examples are being run in a fully-instrumented Axiom
image. The idea is to sort the integrals based on which 'integrate'
function they invoke. This will help categorize the 'integrate'
functions in order to ensure test coverage and improve documentation.

-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Axiom interpreter merge
-=-=-=-=-=-=-=-=-=-=-=-=-=-=
The src/interp directory still contains lisp files that need to be
merged. This week the src/interp/regress.lisp file was merged and
removed.

-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Axiom CATS expansion
-=-=-=-=-=-=-=-=-=-=-=-=-=-=

The Computer Algebra Test Suite (CATS) now contains a section with
graphics tests. These are based on published graphs, such as the
CRC Standard Curves and Surfaces book. More graphs have been
written but not yet pushed.

-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Axiom buglist
-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Bugs in Fricas were tested in Axiom. Those that failed were either
fixed or added to the buglist file. As a result, effort 2 was completed.

All of the Axiom test suite was run against Fricas to uncover fixed
bugs that need to be backported.

-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Axiom makefile removal
-=-=-=-=-=-=-=-=-=-=-=-=-=-=
Portions of the Makefiles have been reworked to simplify the process.
This is still 'in process' and not yet pushed.
books/bookvol5 merge and remove regress.lisp

The regression code is now part of the interpreter