books/bookvol10.3 fixed bug 7297: Extraneous "#\" characters
fixed by 20150126.01.fmp.patch (sent by Frederick H. Pitts)
=========================================================================
bug 7297: Extraneous "#\" characters in Axiom TeX output

    The Axiom TeX output for a number with radix greater than 10
potentially contains "#\A" (A is any upper case alphabetic character)
where it should contain just "A" (the double quotes are mine, they do
not appear in the actual output).  This particular bug is used as a
Axiom debugging example in "Volume 4: Axiom Developers Guide' on page
91, but no fix is presented there.  In the release notes I stumbled over
a fix to essentially the same bug in HTML output.  A quick fix applied
to the TeX case is to replace the invocation of the Lisp function
"object2String" with "mathObject2String" (again, the double quotes are
mine) in the file bookvol10.3.pamphlet, line 132043.

    I came across this issue while trying to use TeXmacs as a front end
to Axiom.  In that environment, the above fix generates nicely formatted
numbers with fixed width characters (i.e. monospaced) and over-lining as
needed.  On the other hand, if I run Axiom in an Emacs shell and attempt
to use latex-math-preview-expression to format the Axiom TeX output, the
number is not so nicely formatted.  The letter digits are slanted and
variable width while the number digits are upright and fixed width.
Maybe a "\rm" font specification command is needed in the Axiom TeX
output for a number with radix greater than 10.

reported on 25 Jan 2015 by Frederick H. Pitts <fred.pitts@comcast.net>