src/doc/s2-us.msgs start inline of message text
This is a startup test for inlining message text and removing s2-us.msgs.

The s2-us.msgs was intended to allow error messages to be internationalized.
NAG did not use this mechanism. Recent discussions makes it obvious that
this mechanism is never going to be used so this effort will remove it.

A call to fetchKeyedMsg now accepts a string directly.
A test case for the compiler is included (bookvol9).
A fix to i-analy which used strings for the key is included.