src/doc/msgs/s2-us.msgs: removed. replace code with msg text
s2-us.msgs was intended to allow messages to be translated from
English to other languages. This was never done. In fact, new
messages added after this file was created were never added to
the message file. As a result this file is not needed. The
messages were inserted into the original source files.