src/interp/interop.lisp merge and purge code
Goal: move toward Literate

The interop.lisp file contained code to handle a prior representation
of Scratchpad code. That representation no longer exists so compiling
the code was a waste of time. Functions that checked for the prior
representation always returned false, wasting runtime.  Some of the
code was used elsewhere and it was moved or merged into the interpreter
book (5), the rest was deleted. The interop.lisp file was removed.