books/bookvol10.* add COQ stanzas
Goal: Proving Axiom Correct

Stanzas were added to the algebra that contain the executable code
with associated signatures. These stanzas are automatically extracted
to the obj/sys/proofs/coq.v file. This file is piped through coqtop
to run the proofs.

Building stanzas is not complete but is (maybe) sufficient to construct
a first proof.