src/etc/asq.c remove compress code
Goal: bug fix

asq is a command line tool to print information from Axiom databases.

Axiom no longer has a compress database as compression for diskettes
is a waste of time and space. Removing the compress database broke asq
which was trying to open it. The asq code has been modified to remove
the compress database code.