books/bookvol10.* standard format, fix documentation
Goal: Axiom Literate Programming

All of the exports of the algebra now follow a common
format. This common format makes the algebra easier to
read by removing variations in style.

Every export is indented by 2 spaces in its enclosing block.

Every export is of the form

  FunctionName : (args) -> (results)
    ++ FunctionName

so that the : is considered separate from the function name
as a syntax element and only one space separates each term.
This seemed to be the most commonly chosen format. It was
applied everywhere so the syntax is uniformly readable.

Unnecessary indentation was removed.

Note that these changes were only applied to the signature
section. Future work will target the constructors and the
implementation sections.

The ++ documentation errors for all algebra files related
to missing left/right parens and improper first word in
comments have been fixed. This fixes bugs 40320, 40361,
40339, 40338, 40336, 40335, 40333, 40326, 40321b, 40319,
40313, 40310, 40308, 40307, 40306, 40305, 40303, 40302,
40295, 40291, 40290, 40289, 40286, 40285, 40284, 40283,
40282, 40281, 40280, 40279, 40278, 40276, 40275, 40272,
40271, 40270, 40269, 40267, 40266, 40265b, 40263, 40262,
40261, 40260, 40259, 40252, 40250, 40249, 40248b, 40247,
40244, 40240, 40239, 40237, 40236, 40230, 40229, 40228,
40227, 40226, 40225, 40223, 40221, 40218, 40217, 40216,
40215, 40214, 40212b, 40209, 40208, 40207, 40206, 40204,
40203, 40201, 40200, 40198, 40197, 40196, 40194, 40191,
40190, 40187, 40186, 40185, 40182, 40181, 40179, 40178,
40175, 40173, 40172, 40171, 40170, 40169, 40168, 40167,
40164, 40162, 40160, 40159, 40156, 40154, 40153, 40152,
40151, 40150, 40149, 40148, 40147, 40146, 40145, 40144,
40143, 40141, 40140, 40139, 40137, 40136, 40135, 40134a,
40133, 40129, 40128, 40126a, 40125, 40124, 40121, 40120b,
40119, 40113, 40110, 40109, 40108, 40105, 40104, 40103,
40102, 40094, 40091, 40090, 40089, 40088, 40087, 40084,
40083, 40082b, 40076, 40075, 40073, 40072, 40062, 40061,
40060, 40051, 40045, 40037b, 40029b, 40028b, 40027,
40026, 40025, 40024, 40023, 40020, 40018, 40017, 40016,
40015, 40014, 40012, 40011, 40010, 40008, 40006, 40005,
40004, 40003, 40002, 40001, 40000, 40318, 40316, 40315,
40314, 40312, 40311, 40309, 40293, 40287, 20572,