src/input/dave89 Davenport Looking at a set of equations
Goal: Axiom Test Suite

These are the equations from Dave89, Davenport's
``Looking at a set of equations'' paper.

"This working paper describes our experiences with using the
Groebner-basis method [Buchberger, 1985] to solve some related systems
of polynomial equations. While we have not yet been able to solve the
system that was our primary motivation, we feel that these experiences
may prove useful to others investigating Buchberger's algorithm in
this context, especially when, as is the case for the system under
investigation, the equations are highly structured. We conclude with
some examples of the polynomials that we factored in the course of
this investigation."