src/axiom-website/documentation.html add quotes
Goal: Axiom Maintenance

The fundamental purpose of scientific discourse is not the mere presentation
of information and thought, but rather its actual communication.
-- George Gopen and Judith Swan

Literate programming ... because yourself from the past does not answer emails.
-- Tim Daly

...design of systems of computer algebra not getting the recognition
they deserve in terms of the object-oriented community. I should say
that the models that Scratchpad and later Axiom had abouy abstraction
are, in my view, superior to all the models that have appeared of
object-orientation.
-- Gaston Gonnet [Gonn05]

I think that that's the big contribution of computer algebra --
making the knowledge of lots of mathematicins that worked for a
long time but have put all their results in journals that are
very difficult to extract and to find, putting it at the fingertips
of everybody. I think that that's one of the most important contributions.
-- Gaston Gonnet [Gonn05]